First pass BRAM output for frame 0
A structure for the current data in BRAM bas been added in the same
display format as the data in the software debug register readout for
direct comparison. This has now been tested by presenting a single frame
worth of data from RAM (8192 bits), and there is a refresh button to
re-read the data in this first frame.

There is still not control beyond this from the BRAM_CTRL IP, and
currently only a single frame is displayed. More control of displaying a
range of or otherwise selectable frames is to be added next.